Added a Picture-in-Picture button and a small bugfix
I was going to create a CR extension just for the PIP feature, but saw that this one existed, so decided to add the feature to an existing extension.
Also, fixed a bug where the extension's  and CR's "SKIP TO" buttons show up together, overlapping. Made only the extension's button show and made it exactly like CR's (basically replacing it).